requirements.txt
<<<<<<< HEAD
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.38        # <-- обновлено!
psycopg2-binary
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic>=2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
=======
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.38        # <-- обновлено!
psycopg2-binary
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic>=2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
>>>>>>> 7a2eea49b24341b939241f3433708164a87b6511
email-validator==2.1.0